+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/index.js":"1","/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/App.js":"2","/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/reportWebVitals.js":"3"},{"size":500,"mtime":1610564199011,"results":"4","hashOfConfig":"5"},{"size":528,"mtime":1610564199009,"results":"6","hashOfConfig":"5"},{"size":362,"mtime":1610564199012,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1rpnamh",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/index.js",[],"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/App.js",[],"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/reportWebVitals.js",[]]
+[{"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/index.js":"1","/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/reportWebVitals.js":"2","/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/Components/App/index.js":"3"},{"size":511,"mtime":1610643334552,"results":"4","hashOfConfig":"5"},{"size":362,"mtime":1610564199012,"results":"6","hashOfConfig":"5"},{"size":439,"mtime":1610643100296,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},"1rpnamh",{"filePath":"11","messages":"12","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"10"},{"filePath":"13","messages":"14","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/index.js",[],["15","16"],"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/reportWebVitals.js",[],"/Users/olgamorgan/turing/3module/projects/final-project/random-coctail-generator/src/Components/App/index.js",[],{"ruleId":"17","replacedBy":"18"},{"ruleId":"19","replacedBy":"20"},"no-native-reassign",["21"],"no-negated-in-lhs",["22"],"no-global-assign","no-unsafe-negation"]
